test(profile): M2-M4 — split oversized test files (all now ≤287 lines)
appeal_notifier_test.dart      353 → 275 (draft group extracted)
appeal_notifier_draft_test.dart        NEW 135 (draft persistence)
appeal_screen_test.dart        345 → 220 (interaction groups extracted)
appeal_screen_interaction_test.dart    NEW 204 (snackbar + discard dialog)
suspension_gate_screen_test.dart 326 → 287 (semantics groups extracted)
suspension_gate_screen_semantics_test.dart NEW 141 (PopScope + liveRegion)
