+++ test/features/profile/presentation/viewmodels/appeal_notifier_test.dart
@@ -271,83 +271,4 @@ void main() {
       expect(state, isA<AsyncError<void>>());
     });
   });
-
-  group('AppealNotifier — draft persistence', () {
-    test('saveDraft stores value under appeal_draft_{id} key', () async {
-      final container = _container(repo: repo, analytics: analytics);
-      addTearDown(container.dispose);
-
-      await container
-          .read(appealNotifierProvider.notifier)
-          .saveDraft(sanctionId: 'sanction-abc', body: 'my draft text');
-
-      final prefs = await SharedPreferences.getInstance();
-      expect(prefs.getString('appeal_draft_sanction-abc'), 'my draft text');
-    });
-
-    test('loadDraft returns null when no draft saved', () async {
-      final container = _container(repo: repo, analytics: analytics);
-      addTearDown(container.dispose);
-
-      final draft = await container
-          .read(appealNotifierProvider.notifier)
-          .loadDraft(sanctionId: 'no-draft-here');
-      expect(draft, isNull);
-    });
-
-    test('loadDraft returns saved draft', () async {
-      SharedPreferences.setMockInitialValues({
-        'appeal_draft_s-2': 'saved draft',
-      });
-
-      final container = _container(repo: repo, analytics: analytics);
-      addTearDown(container.dispose);
-
-      final draft = await container
-          .read(appealNotifierProvider.notifier)
-          .loadDraft(sanctionId: 's-2');
-      expect(draft, 'saved draft');
-    });
-
-    test('clearDraft removes the key', () async {
-      SharedPreferences.setMockInitialValues({'appeal_draft_s-3': 'to clear'});
-
-      final container = _container(repo: repo, analytics: analytics);
-      addTearDown(container.dispose);
-
-      await container
-          .read(appealNotifierProvider.notifier)
-          .clearDraft(sanctionId: 's-3');
-
-      final prefs = await SharedPreferences.getInstance();
-      expect(prefs.getString('appeal_draft_s-3'), isNull);
-    });
-
-    test('saveDraft key format is appeal_draft_<sanctionId>', () async {
-      final container = _container(repo: repo, analytics: analytics);
-      addTearDown(container.dispose);
-
-      await container
-          .read(appealNotifierProvider.notifier)
-          .saveDraft(sanctionId: 'xyz-789', body: 'draft body');
-
-      final prefs = await SharedPreferences.getInstance();
-      expect(prefs.getString('appeal_draft_xyz-789'), 'draft body');
-    });
-
-    test(
-      'saveDraft is idempotent — overwriting with same value is safe',
-      () async {
-        final container = _container(repo: repo, analytics: analytics);
-        addTearDown(container.dispose);
-
-        final notifier = container.read(appealNotifierProvider.notifier);
-        await notifier.saveDraft(sanctionId: 's-4', body: 'same');
-        await notifier.saveDraft(sanctionId: 's-4', body: 'same');
-
-        final prefs = await SharedPreferences.getInstance();
-        expect(prefs.getString('appeal_draft_s-4'), 'same');
-      },
-    );
-  });
 }
